Builds using the ambient Trunk-Recorder files.
The `<>` version of this builds using the system installed Trunk-Recorder files. Meaning that you needed to have already built and installed Trunk-Recorder with `make && make install`. With this, it will pull the "ambient" Trunk-Recorder source files from the parent directory allowing you to make and install Trunk-Recorder for the first time with this plugin in the user_plugins directory.